backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.15",
    "fastapi>=0.127.0",
    "mangum>=0.19.0",
    "openai>=2.14.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
]
